Frontend bug fixing

+++ src/lib/cell/SimulationContext.cpp
@@ -222,7 +222,7 @@ std::vector<Disc> SimulationContext::createDiscGridFromDistribution(const Simula
 
     for (int i = 0; i < simulationConfig.setup.discCount && !discPositions.empty(); ++i)
     {
-        int randomNumber = mathutils::getRandomNumber<double>(0, 1);
+        double randomNumber = mathutils::getRandomNumber<double>(0, 1);
 
         for (const auto& [discType, percentage] : discTypes)
         {
